moving content and options into lib as well
