+++ lib/content.js
@@ -1,4 +1,4 @@
 import 'babel/polyfill';
-import gittenify from './lib/content/gittenify';
+import gittenify from './content/gittenify';
 
 chrome.storage.local.get(null, gittenify);
