+++ lib/options.js
@@ -1,3 +1,3 @@
-import optionsUi from './lib/options/ui';
+import optionsUi from './options/ui';
 
 chrome.storage.local.get(null, optionsUi);
